Fix Delay Bug
Before move IsValidUid() to current place, event always get stocked if top
event was canceled. That is because we need to check the validation of
top, or we will get out the expired check loop owing to invalid event, which is wrong because
we don't need to take care about invalid event.

Also, add some test codes to test other issues.